Feature/perimetre feedback users (#419)
* feat: Affelnet: only one status option for "peut intégrer"

* feat: add confirm on rule delete

* test: add tests on confirm delete

and clean calls to hooks on unmount

+++ ui/src/pages/perimetre/components/RuleModal.jsx
@@ -105,6 +105,7 @@ const RuleModal = ({ isOpen, onClose, rule, onUpdateRule, onDeleteRule, onCreate
     num_academie,
   } = rule ?? {};
 
+  let isClosing = false;
   const isCreating = !rule;
   const initialRef = React.useRef();
   const toast = useToast();
@@ -240,7 +241,8 @@ const RuleModal = ({ isOpen, onClose, rule, onUpdateRule, onDeleteRule, onCreate
   });
 
   const close = () => {
-    resetForm();
+    isClosing = true;
+    // resetForm();
     onClose();
   };
 
@@ -257,10 +259,12 @@ const RuleModal = ({ isOpen, onClose, rule, onUpdateRule, onDeleteRule, onCreate
       setCount(result);
     };
 
-    if (values.niveau && values.diplome) {
-      run();
-    } else {
-      setCount(0);
+    if (!isClosing) {
+      if (values.niveau && values.diplome) {
+        run();
+      } else {
+        setCount(0);
+      }
     }
   }, [values.niveau, values.diplome, values.regle, academie]);
 
@@ -511,17 +515,20 @@ const RuleModal = ({ isOpen, onClose, rule, onUpdateRule, onDeleteRule, onCreate
                       colorScheme="red"
                       borderRadius="none"
                       onClick={async () => {
-                        await onDeleteRule({
-                          _id: idRule,
-                        });
+                        const isUserSure = window.confirm("Voulez vous vraiment supprimer ce diplôme ?");
+                        if (isUserSure) {
+                          await onDeleteRule({
+                            _id: idRule,
+                          });
 
-                        toast({
-                          title: "Suppression",
-                          description: `La règle "${nom_regle_complementaire}" a été supprimée`,
-                          status: "success",
-                          duration: 5000,
-                        });
-                        close();
+                          toast({
+                            title: "Suppression",
+                            description: `La règle "${nom_regle_complementaire}" a été supprimée`,
+                            status: "success",
+                            duration: 5000,
+                          });
+                          close();
+                        }
                       }}
                     >
                       Supprimer
